requirements.txt
streamlit
tensorflow
torch
torchvision
pillow
numpy
